wip pathing update for frontend
