+++ front/src/views/Home.vue
@@ -1,98 +1,78 @@
 <template>
   <div class="relative">
 
-    <div class="task__project__header">
+    <Header
+      :mode="mode"
+      :user="user"
+      :valid-user="validUser" />
 
-      <ProjectInfoName />
+    <router-view />
 
-    </div>
+<!--    <Card v-if="!validUser" class="max-30 intro-text">-->
+<!--      <p class="intro-text__title"> Welcome to MiniTask </p>-->
+<!--      <p class="intro-text__desc1"> A simple, fast, easy and shareable todo list. </p>-->
+<!--      <p class="intro-text__desc2"> To get started, choose a account option from below. </p>-->
 
-    <div class="task__project__header__controls">
+<!--      <div class="flex-row flex-between intro-text__buttons">-->
+<!--        <router-link-->
+<!--          class="text-bold color-fore"-->
+<!--          to="/user/login">-->
+<!--          Login-->
+<!--        </router-link>-->
+<!--        <router-link-->
+<!--          class="text-bold color-fore"-->
+<!--          to="/user/create">-->
+<!--          Create-->
+<!--        </router-link>-->
+<!--        <button class="intro-text__anon"-->
+<!--                @click="createAnonUser">-->
+<!--          Try-->
+<!--        </button>-->
+<!--      </div>-->
 
-      <div class="flex-row">
-        <ProjectTaskSwitch v-model="mode" :is-enabled="validUser" />
+<!--    </Card>-->
 
-        <UserInfoMini :is-enabled="validUser" />
-      </div>
+<!--    <Card v-if="!validUser" class="max-30 intro-text">-->
+<!--      <p class="intro-text__title"> Features: </p>-->
 
-      <InputBar :mode="mode" :is-enabled="validUser" />
+<!--      <ul>-->
+<!--        <li> No need for an account </li>-->
+<!--        <li> A quick overview of progress </li>-->
+<!--        <li> Share projects with others </li>-->
+<!--        <li> Works on Mobile and Desktop </li>-->
+<!--      </ul>-->
 
-    </div>
+<!--    </Card>-->
 
-    <Card v-if="!validUser" class="max-30 intro-text">
-      <p class="intro-text__title"> Welcome to MiniTask </p>
-      <p class="intro-text__desc1"> A simple, fast, easy and shareable todo list. </p>
-      <p class="intro-text__desc2"> To get started, choose a account option from below. </p>
+<!--    <div v-else>-->
+<!--      <ProjectsList v-if="isProjects" @showTasks="showTasks" />-->
 
-      <div class="flex-row flex-between intro-text__buttons">
-        <router-link
-          class="text-bold color-fore"
-          to="/user/login">
-          Login
-        </router-link>
-        <router-link
-          class="text-bold color-fore"
-          to="/user/create">
-          Create
-        </router-link>
-        <button class="intro-text__anon"
-                @click="createAnonUser">
-          Try
-        </button>
-      </div>
+<!--      <TasksList v-if="isTasks" />-->
 
-    </Card>
-
-    <Card v-if="!validUser" class="max-30 intro-text">
-      <p class="intro-text__title"> Features: </p>
-
-      <ul>
-        <li> No need for an account </li>
-        <li> A quick overview of progress </li>
-        <li> Share projects with others </li>
-        <li> Works on Mobile and Desktop </li>
-      </ul>
-
-    </Card>
-
-    <div v-else>
-      <ProjectsList v-if="isProjects" @showTasks="showTasks" />
-
-      <TasksList v-if="isTasks" />
-
-    </div>
+<!--    </div>-->
 
   </div>
 </template>
 
 <script>
+import Header from '../components/Header'
 import modes from '../constants/modes'
-import Card from '../components/general/Card'
-import InputBar from '../components/InputBar'
-import UserInfoMini from '../components/UserInfoMini'
-import TasksList from '../components/TasksList'
-import ProjectsList from '../components/ProjectsList'
-import ProjectInfoName from '../components/ProjectInfoName'
-import ProjectTaskSwitch from '../components/ProjectTaskSwitch'
+// import Card from '../components/general/Card'
+// import TasksList from '../components/TasksList'
+// import ProjectsList from '../components/ProjectsList'
 
 export default {
   name: 'Home',
   components: {
-    Card,
-    InputBar,
-    UserInfoMini,
-    ProjectsList,
-    ProjectInfoName,
-    ProjectTaskSwitch,
-    TasksList
+    Header
+    // Card,
+    // TasksList,
+    // ProjectsList
   },
-  data () {
-    return {
-      mode: modes.PROJECTS,
-      test: {
-        isWaiting: false,
-        isDone: false
-      }
+  props: {
+    mode: {
+      type: String,
+      default: modes.CLEAR
     }
   },
   computed: {
@@ -101,35 +81,19 @@ export default {
     },
     validUser: function () {
       return !(this.user.id === null || this.user.id < 0)
-    },
-    isProjects: function () {
-      return this.mode === modes.PROJECTS
-    },
-    isTasks: function () {
-      return this.mode === modes.TASKS
     }
+    // isProjects: function () {
+    //   return this.mode === modes.PROJECTS
+    // },
+    // isTasks: function () {
+    //   return this.mode === modes.TASKS
+    // }
   },
   methods: {
-    showTasks: function () {
-      this.mode = modes.TASKS
-    },
-    createAnonUser: function () {
-      const userTmp = this.$store.getters['user/user']
-      let initPromise = Promise.resolve(userTmp)
+    // showTasks: function () {
+    //   this.mode = modes.TASKS
+    // },
 
-      if (userTmp && userTmp.id < 0) {
-        initPromise = this.$store.dispatch('user/createAnon')
-      }
-
-      return initPromise
-        .then(user => {
-          return this.$store.dispatch('projects/getProjectsByUserId',
-            { user: user.id })
-        })
-        .catch(err => {
-          this.$store.commit('toasts/toastAdd', err)
-        })
-    }
   }
 }
 </script>
